Change package distribution and refactor Message class
