+++ communitymanager/src/main/java/org/pesmypetcare/communitymanager/managers/GroupManagerClient.java
@@ -11,7 +11,7 @@ import org.pesmypetcare.communitymanager.datacontainers.TagData;
 import org.pesmypetcare.httptools.HttpClient;
 import org.pesmypetcare.httptools.HttpParameter;
 import org.pesmypetcare.httptools.HttpResponse;
-import org.pesmypetcare.httptools.MyPetCareException;
+import org.pesmypetcare.httptools.exceptions.MyPetCareException;
 import org.pesmypetcare.httptools.RequestMethod;
 
 import java.lang.reflect.Type;
