+++ httptools/src/main/java/org/pesmypetcare/httptools/exceptions/PreviousEndDateException.java
@@ -1,0 +1,7 @@
+package org.pesmypetcare.httptools.exceptions;
+
+/**
+ * @author Marc Simó
+ */
+public class PreviousEndDateException extends Exception {
+}
