+++ usermanager/src/main/java/org/pesmypetcare/usermanager/clients/pet/PetManagerClient.java
@@ -9,8 +9,8 @@ import com.google.gson.reflect.TypeToken;
 import org.json.JSONObject;
 import org.pesmypetcare.httptools.HttpClient;
 import org.pesmypetcare.httptools.HttpParameter;
-import org.pesmypetcare.httptools.MyPetCareException;
 import org.pesmypetcare.httptools.RequestMethod;
+import org.pesmypetcare.httptools.exceptions.MyPetCareException;
 import org.pesmypetcare.usermanager.clients.TaskManager;
 import org.pesmypetcare.usermanager.datacontainers.pet.Pet;
 import org.pesmypetcare.usermanager.datacontainers.pet.PetCollectionField;
