+++ usermanager/src/main/java/org/pesmypetcare/usermanager/datacontainers/EventData.java
@@ -4,6 +4,7 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import org.json.JSONObject;
+import org.pesmypetcare.httptools.utilities.DateTime;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
